Put tests in their own module
